src/requirements.txt
Flask==3.0.3
functions-framework==3.8.0
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-core==2.4.1
google-cloud-functions==1.16.4
google-cloud-storage==2.17.0
google-cloud-tasks==2.16.4
googleapis-common-protos==1.63.2
requests==2.32.3
PyJWT==2.8.0
